+++ plugin/src/integrationTest/kotlin/com/vanniktech/maven/publish/ProjectSpecRunner.kt
@@ -1,5 +1,6 @@
 package com.vanniktech.maven.publish
 
+import com.vanniktech.maven.publish.IntegrationTestBuildConfig.QUICK_TEST
 import java.nio.file.Files
 import java.nio.file.Path
 import kotlin.io.path.absolutePathString
@@ -32,8 +33,11 @@ fun ProjectSpec.run(fixtures: Path, temp: Path, options: TestOptions): ProjectRe
     .withProjectDir(project.toFile())
     .withDebug(true)
     .withArguments(arguments)
-    .withTestKitDir(temp.resolve("test-kit-dir").toFile())
-    .build()
+    .apply {
+      if (!QUICK_TEST) {
+        withTestKitDir(temp.resolve("test-kit-dir").toFile())
+      }
+    }.build()
 
   return ProjectResult(
     result = result,